+++ source/android-library/app/src/main/java/com/google/unity/ads/PluginUtils.java
@@ -223,5 +223,4 @@ public class PluginUtils {
                     exception.getLocalizedMessage()));
         }
     }
-
 }