+++ source/plugin/Assets/GoogleMobileAds/Api/AdLoader.cs
@@ -15,6 +15,7 @@
 using System;
 using System.Collections.Generic;
 using System.Reflection;
+using UnityEngine;
 
 using GoogleMobileAds.Common;
 
@@ -105,7 +106,6 @@ namespace GoogleMobileAds.Api
             {
                 get; private set;
             }
-
 
             public Builder ForCustomNativeAd(string templateId)
             {