+++ source/plugin/Assets/GoogleMobileAds/Common/MobileAdsEventExecutor.cs
@@ -22,7 +22,7 @@ namespace GoogleMobileAds.Common
 {
     public class MobileAdsEventExecutor : MonoBehaviour
     {
-        private static MobileAdsEventExecutor instance = null;
+        public static MobileAdsEventExecutor instance = null;
 
         private static List<Action> adEventsQueue = new List<Action>();