+++ source/plugin/Assets/GoogleMobileAds/Common/Utils.cs
@@ -13,6 +13,7 @@
 // limitations under the License.
 
 using System;
+using System.Collections;
 using System.Collections.Generic;
 
 using UnityEngine;
@@ -45,6 +46,5 @@ namespace GoogleMobileAds.Common
 
             return nativeAdTexture;
         }
-
     }
 }