+++ source/plugin/Assets/GoogleMobileAds/Platforms/Android/MobileAdsClient.cs
@@ -14,17 +14,18 @@
 
 #if UNITY_ANDROID
 
+using System;
 using UnityEngine;
 
+using GoogleMobileAds.Api;
 using GoogleMobileAds.Common;
 
 namespace GoogleMobileAds.Android
 {
-    public class MobileAdsClient : IMobileAdsClient
+    public class MobileAdsClient :
+    IMobileAdsClient
     {
         private static MobileAdsClient instance = new MobileAdsClient();
-
-        private MobileAdsClient(){}
 
         public static MobileAdsClient Instance
         {
@@ -42,7 +43,6 @@ namespace GoogleMobileAds.Android
             AndroidJavaClass mobileAdsClass = new AndroidJavaClass(Utils.MobileAdsClassName);
             mobileAdsClass.CallStatic("initialize", activity, appId);
         }
-
         public void SetApplicationVolume(float volume)
         {
             AndroidJavaClass mobileAdsClass = new AndroidJavaClass(Utils.MobileAdsClassName);