+++ source/plugin/Assets/GoogleMobileAds/Platforms/iOS/MobileAdsClient.cs
@@ -14,8 +14,11 @@
 
 #if UNITY_IOS
 
+using System;
+using System.Runtime.InteropServices;
 using UnityEngine;
 
+using GoogleMobileAds.Api;
 using GoogleMobileAds.Common;
 
 namespace GoogleMobileAds.iOS
@@ -23,8 +26,11 @@ namespace GoogleMobileAds.iOS
     public class MobileAdsClient : IMobileAdsClient
     {
         private static MobileAdsClient instance = new MobileAdsClient();
-
-        private MobileAdsClient(){}
+        private IntPtr mobileAdsClientPtr;
+        private MobileAdsClient()
+        {
+            this.mobileAdsClientPtr = (IntPtr)GCHandle.Alloc(this);
+        }
 
         public static MobileAdsClient Instance
         {
@@ -38,7 +44,6 @@ namespace GoogleMobileAds.iOS
         {
             Externs.GADUInitialize(appId);
         }
-
         public void SetApplicationVolume(float volume)
         {
             Externs.GADUSetApplicationVolume(volume);
@@ -52,6 +57,21 @@ namespace GoogleMobileAds.iOS
         public void SetiOSAppPauseOnBackground(bool pause)
         {
             Externs.GADUSetiOSAppPauseOnBackground(pause);
+        }
+        private static MobileAdsClient IntPtrToMobileAdsClient(IntPtr mobileAdsClient)
+        {
+            GCHandle handle = (GCHandle)mobileAdsClient;
+            return handle.Target as MobileAdsClient;
+        }
+
+        public void Dispose()
+        {
+            ((GCHandle)this.mobileAdsClientPtr).Free();
+        }
+
+        ~MobileAdsClient()
+        {
+            this.Dispose();
         }
     }
 }